fix: 유저 신고 UseCase, Repository, DTO, Entity Vote Domain 으로 폴더 이전
