+++ 24th-App-Team-1-iOS/Domain/CommonDomain/Sources/Repositories/CommonRepositoryProtocol.swift
@@ -13,7 +13,6 @@ public protocol CommonRepositoryProtocol {
     func fetchUserProfileItems() -> Single<UserProfileEntity?>
     func createCheckProfanity(body: CreateCheckProfanityRequest) -> Single<Bool>
     func updateUserProfileItem(body: UpdateUserProfileRequest) -> Single<Bool>
-    func createReportUserItem(body: CreateUserReportRequest) -> Single<CreateReportUserEntity?>
     func fetchVoteOptions() -> Single<VoteResponseEntity?>
     func createProfilePresignedURL(query: CreateProfilePresignedURLQuery) -> Single<CreateProfilePresignedURLEntity?>
     func uploadUserProfileImage(_ image: Data, presigendURL: String) -> Single<Bool>
