+++ 24th-App-Team-1-iOS/Domain/VoteDomain/Sources/Entities/CreateUserReportRequest.swift
