+++ 24th-App-Team-1-iOS/Domain/VoteDomain/Sources/UseCases/CreateReportUserUseCaseP.swift
@@ -1,0 +1,8 @@
+//
+//  CreateReportUserUseCaseP.swift
+//  VoteDomain
+//
+//  Created by 김도현 on 5/30/25.
+//
+
+import Foundation
