+++ 24th-App-Team-1-iOS/Service/VoteService/Sources/DataMapping/CreateReportUserResponseDTO.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-import CommonDomain
+import VoteDomain
 
 public struct CreateReportUserResponseDTO: Decodable {
     public let id: Int
