+++ 24th-App-Team-1-iOS/Service/VoteService/Sources/DataMapping/CreateUserReportRequestDTO.swift
